refactor: standardize type and field naming conventions to PascalCase
- Renamed `bitmap_t` to `Bitmap_t` and updated related function signatures.
- Changed `uint32_t Length` to `size_t Length` in the `List_t` structure.
- Renamed `page_policy_t` to `PagePolicy_t` and `page_t` to `Page_t`, updating related functions.
- Removed the `bitmap_test` function from the bitmap implementation.
- Renamed `thread_state_t` to `ThreadState_t` and updated related fields in `TCB_t`.
- Renamed `task_func_t` to `TaskFunc_t` and updated related function signatures.
- Updated `task_create` to use `size_t` for `page_size` instead of `uint16_t`.
- Refactored state field names in `TCB_t` from lowercase to uppercase (e.g., `state` to `State`).
- Updated scheduler and sync logic to use the new `State` field in `TCB_t`.

+++ Core/Data/Inc/list.h
@@ -2,6 +2,7 @@
 #define __LIST_H__
 
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdint.h>
 
 typedef struct ListItem {
@@ -14,7 +15,7 @@ typedef struct ListItem {
 
 // List structure
 typedef struct List {
-    uint32_t Length;    // Number of items in list
+    size_t Length;      // Number of items in list
     ListItem_t *Current;// Current index
     ListItem_t End;     // End marker
 } List_t;
